feat(trpc): move 19 GitHub extended procedures to github-extended.ts
PR #5 Phase B-C: fork の 19 tRPC procedure (`cleanupMissingWorktrees`,
`createGitHubIssue`, `dispatchGitHubWorkflow`, `getCheckJobSteps`,
`getGitHubRepositoryOverview`, `getGitHubWorkflowRuns`, `getJobLogs`,
`getJobStatuses`, `getMissingWorktrees`, `getPullRequestIdentityCandidates`,
`getWorkflowRunJobs`, `replyToPullRequestComment`, `rerunPullRequestChecks`,
`setActiveSyncWorkspace`, `setPullRequestDraftState`,
`setPullRequestThreadResolution`, `updatePullRequestAssignees`,
`updatePullRequestReviewers`, `uploadGitHubIssueAsset`) と専用 helper を
procedures/git-status.ts から github-extended.ts へ移動。

- git-status.ts: 2073 → 383 行 (upstream #3295 に近い形へ)
- github-extended.ts: 新規 1734 行 (fork 19 procedure + 専用 helper + schema)
- main/server 側 typecheck: グリーン
- renderer 側 typecheck: 57 件エラー (Phase D: renderer 書き換えで解消)

`githubSyncService.initialize` のモジュールトップ呼び出しは git-status.ts
に残置、Phase E で整理する。
